backend/pyproject.toml
[project]
name = "agent"
version = "0.0.1"
description = "Starter template for making a new agent LangGraph."
authors = [{ name = "Fauzan Ishlah", email = "fauzanishlah@gmail.com" }]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.9"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "ipykernel>=6.29.5",
    "langchain>=0.3.25",
    "langchain-google-vertexai>=2.0.24",
    "langgraph>=0.2.6",
    "mlflow>=2.22.0",
    "nltk>=3.9.1",
    "passlib>=1.7.4",
    "psycopg-binary>=3.2.9",
    "psycopg>=3.2.9",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
    "sqlmodel>=0.0.24",
    "tensorflow>=2.19.0",
    "python-jose>=3.4.0",
    "pyjwt>=2.10.1",
    "pydantic-settings>=2.9.1",
    "bcrypt>=4.3.0",
    "uuid7>=0.1.0",
]


[project.optional-dependencies]
dev = ["mypy>=1.11.1", "ruff>=0.6.1"]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["langgraph.templates.agent", "agent"]
[tool.setuptools.package-dir]
"langgraph.templates.agent" = "src/agent"
"agent" = "src/agent"


[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.ruff]
lint.select = [
    "E",    # pycodestyle
    "F",    # pyflakes
    "I",    # isort
    "D",    # pydocstyle
    "D401", # First line should be in imperative mood
    "T201",
    "UP",
]
lint.ignore = [
    "UP006",
    "UP007",
    # We actually do want to import from typing_extensions
    "UP035",
    # Relax the convention by _not_ requiring documentation for every function parameter.
    "D417",
    "E501",
]
[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "UP"]
[tool.ruff.lint.pydocstyle]
convention = "google"

[dependency-groups]
dev = [
    "anyio>=4.7.0",
    "langgraph-cli[inmem]>=0.2.8",
    "mypy>=1.13.0",
    "pytest>=8.3.5",
    "ruff>=0.8.2",
]
